requirements.txt
langchain==0.1.14
langchain-community==0.0.37
langchain-openai==0.1.8
faiss-cpu==1.7.4
pypdf==4.1.1
python-dotenv==1.0.0
pydantic==2.5.3
pytest==7.4.4
pytest-asyncio==0.23.3
